moved rootlayout to components
